# specs/001-create-a-mobile/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: FitFusion AI Workout API
  description: REST API for AI-powered workout generation and management with 8 specialized CrewAI agents
  version: 1.0.0
  contact:
    name: FitFusion Development
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://your-app.hf.space
    description: Production server (Huggingface Spaces)
paths:
...
  /api/equipment:
    get:
      summary: List user equipment
      operationId: listEquipment
      responses:
        '200':
          description: Equipment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Equipment'
    
    post:
      summary: Add new equipment
      operationId: addEquipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentCreate'
      responses:
        '201':
          description: Equipment added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
  /api/equipment/{equipment_id}:
    put:
      summary: Update equipment
      operationId: updateEquipment
      parameters:
        - name: equipment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentUpdate'
      responses:
        '200':
          description: Equipment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
    
    delete:
      summary: Remove equipment
      operationId: removeEquipment
      parameters:
        - name: equipment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Equipment removed

  # AI Workout Generation
...
components:
  schemas:
...
    Equipment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
          enum: [weights, cardio, resistance, bodyweight, flexibility]
        specifications:
          type: object
        space_requirements:
          type: object
        noise_characteristics:
          type: object
        condition:
          type: string
          enum: [excellent, good, fair, needs_repair]
        is_available:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...